+++ src/issuingInvoice/log.php
@@ -41,7 +41,7 @@ class Log extends Resource
     /**
     # Retrieve IssuingInvoice\Logs
 
-    Receive a generator of IssuingInvoice\Log objects previously created in the Stark Infra API
+    Receive an enumerator of IssuingInvoice\Log objects previously created in the Stark Infra API
 
     ## Parameters (optional):
         - limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
@@ -52,7 +52,7 @@ class Log extends Resource
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
     ## Return:
-        - generator of IssuingInvoice\Log objects with updated attributes
+        - enumerator of IssuingInvoice\Log objects with updated attributes
      */
     public static function query($options = [], $user = null)
     {