+++ src/pixStatement/pixStatement.php
@@ -4,41 +4,40 @@ namespace StarkInfra;
 use StarkInfra\Utils\Resource;
 use StarkInfra\Utils\Checks;
 use StarkInfra\Utils\Rest;
-use StarkInfra\Utils\StarkDate;
 
 
 class PixStatement extends Resource
 {
     /**
-    # PixStatement  object
+    # PixStatement object
 
     When you initialize a PixStatement , the entity will not be automatically
     created in the Stark Infra API. The 'create' function sends the objects
     to the Stark Infra API and returns the array of created objects.
 
     ## Parameters (required):
-        - after [string]: transactions that happened at this date are stored in the PixStatement, must be the same as before. ex: (2022-01-01)
-        - before [string]: transactions that happened at this date are stored in the PixStatement, must be the same as after. ex: (2022-01-01)
+        - after [Date, Datetime or string]: transactions that happened at this date are stored in the PixStatement, must be the same as before. ex: "2020-04-03"
+        - before [Date, Datetime or string]: transactions that happened at this date are stored in the PixStatement, must be the same as after. ex: "2020-04-03"
         - type [string]: types of entities to include in statement. Options: ["interchange", "interchangeTotal", "transaction"]
 
     ## Attributes (return-only):
-        - id [string, default None]: unique id returned when the PixStatement is created. ex: "5656565656565656"
-        - status [string, default None]: current PixStatement status. ex: "success" or "failed"
-        - transaction_count [integer, default None]: number of transactions that happened during the day that the PixStatement was requested. ex 11
-        - created [string, default None]: creation datetime for the PixStatement. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
-        - updated [string, default None]: latest update datetime for the PixStatement. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
+        - id [string]: unique id returned when the PixStatement is created. ex: "5656565656565656"
+        - status [string]: current PixStatement status. ex: "success" or "failed"
+        - transactionCount [integer]: number of transactions that happened during the day that the PixStatement was requested. ex 11
+        - created [DateTime or string]: creation datetime for the PixStatement. 
+        - updated [DateTime or string]: latest update datetime for the PixStatement. 
      */
     function __construct(array $params)
     {
         parent::__construct($params);
 
-        $this->after = Checks::checkParam($params, "after");
-        $this->before = Checks::checkParam($params, "before");
-        $this->type = Checks::checkParam($params, "type");
-        $this->status = Checks::checkParam($params, "status");
-        $this->transactionCount = Checks::checkParam($params, "transactionCount");
-        $this->created = Checks::checkDateTime(Checks::checkParam($params, "created"));
-        $this->updated = Checks::checkDateTime(Checks::checkParam($params, "updated"));
+        $this-> after = Checks::checkParam($params, "after");
+        $this-> before = Checks::checkParam($params, "before");
+        $this-> type = Checks::checkParam($params, "type");
+        $this-> status = Checks::checkParam($params, "status");
+        $this-> transactionCount = Checks::checkParam($params, "transactionCount");
+        $this-> created = Checks::checkDateTime(Checks::checkParam($params, "created"));
+        $this-> updated = Checks::checkDateTime(Checks::checkParam($params, "updated"));
 
         Checks::checkParams($params);
     }
@@ -50,6 +49,7 @@ class PixStatement extends Resource
 
     ## Parameters (required):
         - statement [PixStatement object]: PixStatement object to be created in the API.
+
     ## Parameters (optional):
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
@@ -58,7 +58,7 @@ class PixStatement extends Resource
      */
     public static function create($statement, $user = null)
     {
-        return Rest::postSingle($user, PixStatement ::resource(), $statement);
+        return Rest::postSingle($user, PixStatement::resource(), $statement);
     }
 
     /**
@@ -86,7 +86,7 @@ class PixStatement extends Resource
     Receive an enumerator of PixStatement  objects previously created in the Stark Infra API
 
     ## Parameters (optional):
-        - limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
+        - limit [integer, default 100]: maximum number of objects to be retrieved. ex: 35
         - ids [array of strings, default null]: array of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
@@ -106,13 +106,13 @@ class PixStatement extends Resource
 
     ## Parameters (optional):
         - cursor [string, default null]: cursor returned on the previous page function call
-        - limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
+        - limit [integer, default 100]: maximum number of objects to be retrieved. ex: 35
         - ids [array of strings, default null]: array of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
     ## Return:
-    - list of PixStatement objects with updated attributes
-    - cursor to retrieve the next page of PixStatement  objects
+        - list of PixStatement objects with updated attributes
+        - cursor to retrieve the next page of PixStatement  objects
      */
     public static function page($options = [], $user = null)
     {
@@ -120,18 +120,18 @@ class PixStatement extends Resource
     }
 
     /**
-    # Retrieve a specific PixStatement
+    # Retrieve a .csv PixStatement
 
-    Receive a single PixStatement  object previously created in the Stark Infra API by passing its id
+    Retrieve a specific PixStatement by its ID in a .csv file.
 
     ## Parameters (required):
-    - id [string]: object unique id. ex: "5656565656565656"
+        - id [string]: object unique id. ex: "5656565656565656"
 
     ## Parameters (optional):
-    - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
+        - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
     ## Return:
-    - PixStatement .zip file containing the csv.
+        - .zip file containing a PixStatement in .csv format
      */
     public static function csv($id, $user = null)
     {
@@ -144,7 +144,7 @@ class PixStatement extends Resource
             return new PixStatement ($array);
         };
         return [
-            "name" => "PixStatement ",
+            "name" => "PixStatement",
             "maker" => $statement,
         ];
     }